--- Archivo de prueba.txt
@@ -1,2 +1,0 @@
-Actualizado de Enrique
-hola
